fix: circular dependencies in Modal & Input (#31817)
* fix(input): fix circular dependencies

* fix(input): fix circular dependencies

* fix(modal): fix circular dependencies

* fix(modal): fix test
