+++ components/input/ClearableLabeledInput.tsx
@@ -2,16 +2,13 @@ import * as React from 'react';
 import classNames from 'classnames';
 import CloseCircleFilled from '@ant-design/icons/CloseCircleFilled';
 import { tuple } from '../_util/type';
-import { InputProps, getInputClassName } from './Input';
+import type { InputProps } from './Input';
 import { DirectionType } from '../config-provider';
 import { SizeType } from '../config-provider/SizeContext';
 import { cloneElement } from '../_util/reactNode';
+import { getInputClassName, hasPrefixSuffix } from './utils';
 
 const ClearableInputType = tuple('text', 'input');
-
-export function hasPrefixSuffix(props: InputProps | ClearableInputProps) {
-  return !!(props.prefix || props.suffix || props.allowClear);
-}
 
 function hasAddon(props: InputProps | ClearableInputProps) {
   return !!(props.addonBefore || props.addonAfter);
@@ -35,7 +32,7 @@ interface BasicProps {
 }
 
 /** This props only for input. */
-interface ClearableInputProps extends BasicProps {
+export interface ClearableInputProps extends BasicProps {
   size?: SizeType;
   suffix?: React.ReactNode;
   prefix?: React.ReactNode;
