+++ components/input/utils.ts
@@ -1,0 +1,25 @@
+import classNames from 'classnames';
+import type { DirectionType } from '../config-provider';
+import type { SizeType } from '../config-provider/SizeContext';
+import type { ClearableInputProps } from './ClearableLabeledInput';
+import type { InputProps } from './Input';
+
+export function getInputClassName(
+  prefixCls: string,
+  bordered: boolean,
+  size?: SizeType,
+  disabled?: boolean,
+  direction?: DirectionType,
+) {
+  return classNames(prefixCls, {
+    [`${prefixCls}-sm`]: size === 'small',
+    [`${prefixCls}-lg`]: size === 'large',
+    [`${prefixCls}-disabled`]: disabled,
+    [`${prefixCls}-rtl`]: direction === 'rtl',
+    [`${prefixCls}-borderless`]: !bordered,
+  });
+}
+
+export function hasPrefixSuffix(props: InputProps | ClearableInputProps) {
+  return !!(props.prefix || props.suffix || props.allowClear);
+}
