+++ components/modal/destroyFns.ts
@@ -1,0 +1,2 @@
+const destroyFns: Array<() => void> = [];
+export default destroyFns;
