+++ components/modal/index.tsx
@@ -1,4 +1,4 @@
-import OriginModal, { ModalFuncProps, destroyFns } from './Modal';
+import OriginModal, { ModalFuncProps } from './Modal';
 import confirm, {
   withWarn,
   withInfo,
@@ -8,6 +8,8 @@ import confirm, {
   ModalStaticFunctions,
   modalGlobalConfig,
 } from './confirm';
+import useModal from './useModal';
+import destroyFns from './destroyFns';
 
 export { ActionButtonProps } from './ActionButton';
 export { ModalProps, ModalFuncProps } from './Modal';
@@ -17,9 +19,15 @@ function modalWarn(props: ModalFuncProps) {
 }
 
 type ModalType = typeof OriginModal &
-  ModalStaticFunctions & { destroyAll: () => void; config: typeof modalGlobalConfig };
+  ModalStaticFunctions & {
+    useModal: typeof useModal;
+    destroyAll: () => void;
+    config: typeof modalGlobalConfig;
+  };
 
 const Modal = OriginModal as ModalType;
+
+Modal.useModal = useModal;
 
 Modal.info = function infoFn(props: ModalFuncProps) {
   return confirm(withInfo(props));
